Move `change_out_bias` to base atomic model
